Asm refactoring (#41)
* rename Interfaces to ASMLib

* patient to monggo entity

* patients store

* new controller

* mongo store and influence and patient services

* new route points

* influences methods for mongo

* complete influences controller

* Update InfluencesDataService.cs

* Update InfluencesDataService.cs

* Update InfluencesDataService.cs

* patients service

* patient controller and service (complete)

* remove patientshandler and postgresql db

* agents api new entities for influences and patient

* agents api refactoring

* commands refactoring

* agents api refactoring

* query refactoring

* refactoring

* Update Program.cs

* refactoring

* patient parameters refactoring

* requests new api

* mongo connection settings

* rename

* Update GetAgeCommand.cs

* string agent type and agentType into agentKey

* rename agent settings

* refactoring settings store

* agent settings store

* init sets request refactoring

* agent type into prediction request

* Update PredictionRequest.cs

* Update SettingsService.cs

* Update PatientsDataService.cs

* remove temp model settings, model meta as mongo document

* agent command api  - mlModel id

* docker-compose refactoring

* Update docker-compose.yml

* env file

* make file

* docker nuild fix

* app sets fixs

* fixes for nullable requests

* Update GetBioageByFuncParamsCommand.cs

* rename patientsService and local store patientsInfo

* isRefresh option to patient request

* predict command

* agents command refactoring

* mongo settings

* age into meta

* Update InternalMetaStorageService.cs

* Delete HomeController.cs

* remove not ness service and job

* api ref

* Update Program.cs

* Update Program.cs

* appsets refactoring

* Update Patient.cs

* Update PatientsDataService.cs

* influence add refactoring

* addPatientParameters

* frombody

* Update PatientsDataService.cs

* Update PredictionRequestsService.cs

* remove views

* Update PredictionRequestsService.cs

* build fixes

* metas init fix

* getpatients route point

* Update PatientsResolver.API.Service.csproj

+++ Interfaces/IPredictRequest.cs
@@ -12,6 +12,6 @@ namespace Interfaces
 
         public string Version { get; set; }
 
-        public float[] Input { get; set; }
+        public double[] Input { get; set; }
     }
 }
